requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.10
requests==2.32.3
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
sqlalchemy==2.0.38
pymysql==1.1.1
pyodbc==5.2.0
cx_Oracle==8.3.0
boto3==1.37.0
loguru==0.7.3
ydata-profiling==4.12.2
setuptools==75.8.0
